pyproject.toml
[project]
name = "api-servic"
version = "0.1.0"
description = ""
authors = [
    {name = "Dima",email = "nubile4446@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.12,<3.15"
dependencies = [
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pytest (>=9.0.1,<10.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "black (>=25.11.0,<26.0.0)",
    "django (>=6.0,<7.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "isort (>=7.0.0,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [
    { include = "core" },
    { include = "questions" },
    { include = "answers" }
]
package-mode = false
